requirements.txt
django==3.2.1
social-auth-app-django==4.0.0
djangorestframework==3.12.4
pillow==8.2.0